Prediction method for itemgn objects

Obtains gamma-values on a test dataset based on an itemgn object from link{cfsens_mgn}.

# S3 method for itemgn
predict(
  obj,
  X_test,
  Y1_test = NULL,
  Y0_test = NULL,
  type = c("ate", "att", "atc"),
  Gamma_max = 5,
  gamma_length = 51
)

Arguments

obj

an object of class itemgn.

X_test

testing covariates.

Y1_test

the potential outcome when treated. The default is NULL. See details.

Y0_test

the potential outcome when not treated. The default is NULL. See details.

type

the type of inference target. Takes value in {"ate", "att", "atc"}. See details.

Gamma_max

the maximum value of Gamma to be considered for sensitivity analysis. The default is 5.

gamma_length

the number of Gamma to be considered for sensitivity analysis. The default is 51.

Value

a vector of gamma-values.

Details

When type = "ate", the inference is valid unconditionally; when type = "att", the inference is valid conditional on T=1, and Y1_test should be provided; when type = "atc", the inference is valid conditional on T=0, and Y0_test should be provided.